+++ lib/deep_cover/custom_requirer.rb
@@ -8,6 +8,12 @@ module DeepCover
       @loaded_features = loaded_features
       @filter = filter
       @paths_being_required = Set.new
+
+      # A Set of the loaded_features for faster access
+      @loaded_features_set = Set.new
+      # A dup of the loaded_features as they are expected to be for the Set to be valid
+      # If this is different from loaded_features, the set should be refreshed
+      @duped_loaded_features_used_for_set = []
     end
 
     # Returns a path to an existing file or nil if none can be found.
@@ -28,11 +34,12 @@ module DeepCover
 
       paths_with_ext = extensions_to_try.map { |ext| path + ext }
 
+      refresh_loaded_features_set
+
       # Doing this check in every case instead of only for absolute_path because ruby has some
       # built-in $LOADED_FEATURES which aren't an absolute path. Ex: enumerator.so, thread.rb
-      paths_with_ext.each do |path_with_ext|
-        return path_with_ext if @loaded_features.include?(path_with_ext)
-      end
+      path_from_loaded_features = first_path_from_loaded_features_set(paths_with_ext)
+      return path_from_loaded_features if path_from_loaded_features
 
       if path == abs_path
         paths_with_ext.each do |path_with_ext|
@@ -40,9 +47,8 @@ module DeepCover
         end
       else
         possible_paths = paths_with_load_paths(paths_with_ext)
-        possible_paths.each do |possible_path|
-          return possible_path if @loaded_features.include?(possible_path)
-        end
+        path_from_loaded_features = first_path_from_loaded_features_set(possible_paths)
+        return path_from_loaded_features if path_from_loaded_features
 
         possible_paths.each do |possible_path|
           next unless File.exist?(possible_path)
@@ -78,10 +84,10 @@ module DeepCover
       end
 
       DeepCover.autoload_tracker.wrap_require(path, found_path) do
-        # Either a problem with resolve_path, or a gem that will be added to the load_path by RubyGems
-        return yield(:not_found) unless found_path
-
         begin
+          # Either a problem with resolve_path, or a gem that will be added to the load_path by RubyGems
+          return yield(:not_found) unless found_path
+
           @paths_being_required.add(found_path)
           return yield(:not_in_covered_paths) unless DeepCover.within_lookup_paths?(found_path)
           return yield(:not_supported) if found_path.end_with?('.so')
@@ -92,6 +98,7 @@ module DeepCover
           @loaded_features << found_path
         ensure
           @paths_being_required.delete(found_path)
+          add_last_loaded_feature_to_set
         end
       end
       true
@@ -126,6 +133,32 @@ module DeepCover
     end
 
     protected
+
+    # updates the loaded_features_set if it needs it
+    def refresh_loaded_features_set
+      return if @duped_loaded_features_used_for_set == @loaded_features
+
+      @duped_loaded_features_used_for_set = @loaded_features.dup
+      @loaded_features_set = Set.new(@duped_loaded_features_used_for_set)
+    end
+
+    # Returns the first path found in the loaded_features_set
+    # Should be called after doing a #refresh_loaded_features_set
+    def first_path_from_loaded_features_set(paths)
+      paths.detect { |path| @loaded_features_set.include?(path) }
+    end
+
+    # Called after a require, adds the last entry of loaded_features to the
+    # loaded_features_set and the clone used to check for a need to refresh
+    # the loaded_features_set. Doing this allows us to never need to update
+    # the loaded_feature_set from scratch (almost? this is a safety precaution)
+    def add_last_loaded_feature_to_set
+      loaded_feature = @loaded_features.last
+      unless @loaded_features_set.include?(loaded_feature)
+        @duped_loaded_features_used_for_set << loaded_feature
+        @loaded_features_set << loaded_feature
+      end
+    end
 
     def paths_with_load_paths(paths)
       paths.flat_map do |path|